+++ scoring_function.h
@@ -61,9 +61,9 @@ long double score(NTSparseGraph& g, const CombinatoricUtility& comb_util,
                   Coloring<Edge,EdgeHash>& editable_edge_orbit_coloring,
                   const std::unordered_set<Edge,EdgeHash>& edge_additions,
                   const std::unordered_set<Edge,EdgeHash>& edge_removals,
-                  const double log2_p_plus, const double log2_p_minus,
-                  const double log2_1_minus_p_plus,
-                  const double log2_1_minus_p_minus);
+                  const long double log2_p_plus, const long double log2_p_minus,
+                  const long double log2_1_minus_p_plus,
+                  const long double log2_1_minus_p_minus);
 
 long double score(NTSparseGraph& g, const CombinatoricUtility& comb_util,
                   const Coloring<int>& node_orbit_coloring,
